### 2020-10-09 #2738 @whedon
[PRE REVIEW]: mcbette: model comparison using babette
**Submitting author:** @richelbilderbeek (<a href="http://orcid.org/0000-0003-1107-7049">Richel Bilderbeek</a>)
**Repository:** <a href="https://github.com/ropensci/mcbette" target ="_blank">https://github.com/ropensci/mcbette</a>
**Version:** v1.8.3
