### 2020-10-09 #2738 @kthyng
Hi @richelbilderbeek! Thanks for your submission. Your submission is missing a few items or needs editing:

* open source license
* DOI's are missing from references (see list above)
* Your paper needs a high level non-technical short introduction to orient a reader who might be interested in using your software to decide whether it is relevant or not.
* you should be able to resize your giant logo in the paper too
