### 2020-10-09 #2738 @richelbilderbeek
This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/360, @ropensci/editors
