### 2020-10-09 ropensci/mcbette#37 @richelbilderbeek
Process feedback JOSS
My JOSS submission can be found at [https://github.com/openjournals/joss-reviews/issues/2738](https://github.com/openjournals/joss-reviews/issues/2738).

Process the first feedback:

> open source license

Upon your request, I have added the LICENSE file.

> DOI's are missing from references (see list above)

Fixed the DOIs indited by [this](https://github.com/openjournals/joss-reviews/issues/2738#issuecomment-706222339) @whelan commit.

> Your paper needs a high level non-technical short introduction to orient a reader who might be interested in using your software to decide whether it is relevant or not.

AFAICS the Introduction serves that purpose:

>``mcbette`` is an R package to do model comparison between
a set of evolutionary models on a DNA alignment, which allows
to select that model that is closest to the process that underlied
the DNA alignment and species tree.

I assume the rOpenSci reviewers agree. If you disagree, let me know and I will make it less technical :+1:

> you should be able to resize your giant logo in the paper too

Done!
