### 2020-10-09 #2738 @kthyng
> This package has been reviewed by rOpenSci: [ropensci/software-review#360](https://github.com/ropensci/software-review/issues/360), @ropensci/editors

@richelbilderbeek ah! I didn't realize that. I will process this submission then.

@marcosvital Please disregard the invitation above.
