### 2020-10-10 ropensci/mcbette#37 @kthyng
>
> The package does have an open source license, which is GPL-3, as can be seen [in this line in the DESCRIPTION file](https://github.com/ropensci/mcbette/blob/master/DESCRIPTION#L31)

You need an actual LICENSE file that contains the license language, instead of the name of an open source license listed as is presently.

>
> > Your paper needs a high level non-technical short introduction to orient a reader who might be interested in using your software to decide whether it is relevant or not.
>
> I assume the rOpenSci review [here](https://github.com/ropensci/software-review/issues/360) is sufficient?
>

I'm not sure what part of that review you are referring to. I mean you should add a few sentences to the start of your JOSS paper that describe your software at a high level. If you have already written some description that would work, that is fine to use.
