### 2020-10-16 ropensci/mcbette#37 @richelbilderbeek
Done, processed the feedback and posted the reply at the original post at [https://github.com/openjournals/joss-reviews/issues/2738](https://github.com/openjournals/joss-reviews/issues/2738)
