### 2020-10-19 #2762 @whedon
[REVIEW]: mcbette: model comparison using babette
**Submitting author:** @richelbilderbeek (<a href="http://orcid.org/0000-0003-1107-7049">Richel Bilderbeek</a>)
**Repository:** <a href="https://github.com/ropensci/mcbette" target ="_blank">https://github.com/ropensci/mcbette</a>
**Version:** v1.8.3
