### 2020-10-19 #2762 @kthyng
This submission was reviewed in [rOpenSci](https://github.com/ropensci/software-review/issues/360).
