### 2020-10-19 #2762 @kthyng
I have some edits for your paper:

- [ ] caption should probably end with a period
- [ ] "allows to select that" --> "allows for selecting the"
- [ ] "underlied" --> "led to"? or "consistent with"? I don't think underlied is a word but I'm not sure what the correct replacement is in this technical sentence.
- [ ] "mcbette can be both be" --> remove a "be"
- [ ] "is aimed to be used" --> "is aimed at being used"
- [ ] "consist out of" --> "consist of"

Also, the capitalization in your references is off. For example, the first reference has "r" in it instead of "R". You can add {} around words to preserve capitalization. Please carefully check all of your references and let me know if you have questions.
