### 2020-10-19 #2762 @kthyng
I see your Zenodo archive at 10.5281/zenodo.4076183. Can you check the author list and metadata to exactly match your JOSS paper?
