### 2020-10-20 1849eec
Use 'INSTALL_BEAST2' in Travis script
