+++ .travis.yml
@@ -14,11 +14,10 @@ matrix:
   include:
   - r: release
     os: linux
+    env: INSTALL_BEAST2=true
   - r: devel
     os: linux
-  # No OSX, due to 'sudo tlmgr update --self' taking too long, see https://travis-ci.org/github/ropensci/mcbette/jobs/729580427#L109
-  #- r: release
-  #  os: osx
+    env: INSTALL_BEAST2=false
 
 # It may be that this package uninstall BEAST2 and/or the packages
 # before and after the tests.
@@ -34,8 +33,6 @@ r_github_packages:
   - jimhester/covr
   - jimhester/lintr
   - MangoTheCat/goodpractice
-  #- ropensci/tracerer
-  #- ropensci/mauricer
 
 before_install:
   # Fix rJava error, from https://travis-ci.community/t/cannot-connect-to-java-from-r/9754/8
@@ -51,10 +48,10 @@ before_install:
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then sudo tlmgr update --self; fi
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then sudo tlmgr install inconsolata; fi
 
-
 before_script:
-  - # echo "mcbette should build with and without BEAST2 installed"
-  - Rscript -e "beastier::install_beast2()"
+  # Install BEAST2. Note: beastier must work without BEAST2 installed as well
+  - if [[ "$INSTALL_BEAST2" == true ]]; then Rscript -e "beastier::install_beast2()"; fi
+  # Check state before mcbette is checked
   - Rscript -e "beastier::is_beast2_installed()"
   - Rscript -e 'mauricer::install_beast2_pkg("NS")'
   - Rscript -e 'mauricer::is_beast2_ns_pkg_installed()'
@@ -66,21 +63,24 @@ before_script:
 script:
   - travis_wait 30 R CMD build .
   - travis_wait 40 R CMD check *tar.gz
+  # Check state after mcbette is checked, which must be the same
   - beast2_installed_after=$(Rscript -e "beastier::is_beast2_installed()")
   - beast2_ns_pkg_installed_after=$(Rscript -e 'mauricer::is_beast2_ns_pkg_installed()')
   - echo $beast2_installed_after
   - echo $beast2_ns_pkg_installed_after
   - if [ "$beast2_installed_before" != "$beast2_installed_after" ]; then echo "ERROR! THEY DIFFER!"; exit 1; fi
   - if [ "$beast2_ns_pkg_installed_before" != "$beast2_ns_pkg_installed_after" ]; then echo "ERROR! THEY DIFFER!"; exit 1; fi
+  # lintr
+  - if [[ "$INSTALL_BEAST2" == false ]]; then Rscript -e 'lintr::expect_lint_free()'; fi
   # Convert the paper to PDF
   - pandoc paper.md -o paper.pdf
   - if [ ! -f paper.pdf ]; then echo "file 'paper.pdf' not found" ; exit 1 ; fi
 
 after_success:
   # Codecov
-  - if [ "$TRAVIS_OS_NAME" == "linux" ] && [ "$TRAVIS_R_VERSION_STRING" == "devel" ]; then travis_wait 60 Rscript -e 'covr::codecov()'; fi
+  - if [[ "$INSTALL_BEAST2" == true ]]; then travis_wait 60 Rscript -e 'covr::codecov()'; fi
   # Goodpractice
-  - if [ "$TRAVIS_OS_NAME" == "linux" ] && [ "$TRAVIS_R_VERSION_STRING" == "release" ]; then travis_wait 60 Rscript -e 'goodpractice::gp()'; fi
[…]
