### 2020-10-20 e4c4651
Add 'INSTALL_BEAST2_NS' in build matrix
