### 2020-10-20 97d3269
Caption ends with period, thank @kthyng, progress https://github.com/openjournals/joss-reviews/issues/2762
