### 2020-10-20 d443b68
Improve choice of words, thank @kthyng, progress https://github.com/openjournals/joss-reviews/issues/2762
