### 2020-10-20 065fc1f
Fix capitalization of references, thank @kthyng, progress https://github.com/openjournals/joss-reviews/issues/2762
