+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Bilderbeek:2018,
-  title={babette: BEAUti 2, BEAST2 and Tracer for R},
+  title={{b}abette: {BEAU}ti 2, {BEAST}2 and {T}racer for {R}},
   author={Bilderbeek, Richèl JC and Etienne, Rampal S},
   journal={Methods in Ecology and Evolution},
   volume={9},
@@ -11,7 +11,7 @@
 }
 
 @article{Bouckaert:2019,
-  title={BEAST 2.5: An advanced software platform for Bayesian evolutionary analysis},
+  title={{BEAST} 2.5: An advanced software platform for {B}ayesian evolutionary analysis},
   author={Bouckaert, Remco and Vaughan, Timothy G and Barido-Sottani, Joelle and Duch{\^e}ne, Sebasti{\'a}n and Fourment, Mathieu and Gavryushkina, Alexandra and Heled, Joseph and Jones, Graham and K{\"u}hnert, Denise and De Maio, Nicola and others},
   journal={PLoS computational biology},
   volume={15},
