### 2020-10-20 ad1deb8
Caption ends with period, thank @kthyng, progress https://github.com/openjournals/joss-reviews/issues/2762
