### 2020-10-20 2ffae4f
Can use BEAST2 installed at a different location
