+++ R/check_beast2_ns_pkg.R
@@ -5,7 +5,7 @@
 #' @export
 check_beast2_ns_pkg <- function() {
   beastier::check_beast2()
-  if (!is_beast2_ns_pkg_installed()) {
+  if (!mauricer::is_beast2_ns_pkg_installed()) {
     stop(
       "BEAST2 'NS' package not installed. ",
       "Tip: use mauricer::install_beast2_pkg(\"NS\")"
