### 2020-10-20 e03974d
beastier v2.2 is on CRAN
