### 2020-10-22 ea74888
Use correct function arguments
