### 2020-10-22 d5ba650
Document all function arguments of 'mcbette_report'
