### 2020-10-22 67ba643
Depend on newer beastier and mauricer packages
