+++ .travis.yml
@@ -39,6 +39,8 @@ env:
     - beast2_ns_pkg_installed_after=-1
 
 r_github_packages:
+  - ropensci/beastier
+  - ropensci/mauricer
   - jimhester/covr
   - jimhester/lintr
   - MangoTheCat/goodpractice
