+++ man/default_params_doc.Rd
@@ -8,6 +8,7 @@ It is intended to inherit function argument documentation.}
 \usage{
 default_params_doc(
   beast2_bin_path,
+  beast2_folder,
   beast2_working_dir,
   beast2_options,
   beast2_optionses,
@@ -31,6 +32,16 @@ default_params_doc(
 }
 \arguments{
 \item{beast2_bin_path}{path to the the BEAST2 binary file}
+
+\item{beast2_folder}{the folder where the BEAST2 is installed.
+Note that this is not the folder where the BEAST2 executable is installed:
+the BEAST2 executable is in a subfolder.
+Use \link[beastier]{get_default_beast2_folder}
+  to get the default BEAST2 folder.
+Use \link[beastier]{get_default_beast2_bin_path}
+  to get the full path to the default BEAST2 executable.
+Use \link[beastier]{get_default_beast2_jar_path}
+  to get the full path to the default BEAST2 jar file.}
 
 \item{beast2_working_dir}{folder in which BEAST2 will run and
 produce intermediate files.
