### 2020-10-22 #2762 @richelbilderbeek
> I see your Zenodo archive at 10.5281/zenodo.4076183. Can you check the author list and metadata to exactly match your JOSS paper?

Done.

@kthyng: AFAICS I have processed your feedback. Thanks for checking up on me! If I missed something, please let me know :+1:
