### 2020-10-22 d037809
Fix style, thanks @lintr-bota
