+++ R/interpret_marg_lik_estimates.R
@@ -23,7 +23,7 @@ interpret_marg_lik_estimates <- function(marg_liks) {
   df_ordered <- stats::na.omit(df)
 
   # Order from high to low
-  df_ordered <- df_ordered[ order(-df_ordered$weight), ]
+  df_ordered <- df_ordered[order(-df_ordered$weight), ]
 
   # Show most convincing first
   cat(" ", sep = "\n")
