+++ R/get_mcbette_state.R
@@ -1,4 +1,5 @@
 #' Get the current state of \link{mcbette}
+#' @inheritParams default_params_doc
 #' @return a \link{list} with the following elements:
 #' \itemize{
 #'   \item{
@@ -16,14 +17,18 @@
 #' @examples
 #' get_mcbette_state()
 #' @export
-get_mcbette_state <- function() {
+get_mcbette_state <- function(
+  beast2_folder = beastier::get_default_beast2_folder()
+) {
   state <- list(
     beast2_installed = NA,
     ns_installed = NA
   )
   if (beastier::is_beast2_installed()) {
     state$beast2_installed <- TRUE
-    state$ns_installed <- mauricer::is_beast2_ns_pkg_installed()
+    state$ns_installed <- mauricer::is_beast2_ns_pkg_installed(
+      beast2_folder = beast2_folder
+    )
   } else {
     state$beast2_installed <- FALSE
   }
