### 2020-10-22 #2762 @kthyng
Is there a new DOI? I checked the same one (10.5281/zenodo.4076184) and it looks the same. I am looking for the title to be "mcbette: model comparison using babette".
