### 2020-10-22 25ee26f
BEAST2 has NS pre-installed
