+++ R/mcbette_self_test.R
@@ -4,10 +4,10 @@
 mcbette_self_test <- function(
   beast2_folder = beastier::get_default_beast2_folder()
 ) {
-  est_marg_lik(
+  mauricer::est_marg_lik(
     fasta_filename = system.file("extdata", "simple.fas", package = "mcbette"),
     inference_model = beautier::create_test_ns_inference_model(),
-    beast2_options = create_beast2_options(
+    beast2_options = beastier::create_beast2_options(
       beast2_path = beastier::get_default_beast2_bin_path(
         beast2_folder = beast2_folder
       )
