### 2020-10-23 510ecc7
Fix style, thanks @lintr-bot
