### 2020-10-24 5b446a1
Simplify 'set_mcbette_state', use correct namespace
