### 2020-10-24 4dd389a
Fix style, thanks @lintr-bot
