### 2020-10-26 fb396bb
Install BEAST2 if it is not installed (instead of installing it when it is already there)
