### 2020-10-26 05549c3
'set_mcbette_state' can uninstall BEAST2
