+++ R/set_mcbette_state.R
@@ -62,4 +62,15 @@ set_mcbette_state <- function(
       verbose = verbose
     )
   }
+
+  # Uninstall BEAST2 if requested
+  if (isFALSE(mcbette_state$beast2_installed) &&
+    beastier::is_beast2_installed(folder_name = beast2_folder)
+  ) {
+    beastier::uninstall_beast2(
+      folder_name = beast2_folder,
+      verbose = verbose
+    )
+  }
+
 }
