### 2020-10-26 0de5c7e
'after' can be NA
