### 2020-10-26 efc28e3
Trailing blank lines are superfluous, thanks @lintr-bot
