### 2020-10-26 cbd6166
Do not test state before and after tests
